tux3: Clean UBSAN warnings for libklib/
The left shift of signed type in libklib/bitops/fls.h.

fls.h:31:5: runtime error: left shift of 666894336 by 2 places cannot be represented in type 'int'

And another one to suppress the following UBSAN warning,

be_byteshift.h:14:14: runtime error: left shift of 195 by 24 places cannot be represented in type 'int'

cast to (u32).

+++ fs/tux3/user/libklib/unaligned/be_byteshift.h
@@ -11,7 +11,7 @@ static inline u16 __get_unaligned_be16(const u8 *p)
 
 static inline u32 __get_unaligned_be32(const u8 *p)
 {
-	return p[0] << 24 | p[1] << 16 | p[2] << 8 | p[3];
+	return (u32)p[0] << 24 | p[1] << 16 | p[2] << 8 | p[3];
 }
 
 static inline u64 __get_unaligned_be64(const u8 *p)
